put the firebase fetch data in ParkingStatus; use strict equal in status children components
