+++ src/ParkingStatus.js
@@ -3,6 +3,7 @@ import styled from 'styled-components'
 import StatusTable from './StatusTable'
 import StatusSummary from './StatusSummary'
 import StatusLegend from './StatusLegend'
+import firebaseApp from './firebase'
 
 const Wrapper = styled.div`
   background-color: rgba(0, 0, 0, 0.6);  // Black with Transparency of 40%
@@ -30,35 +31,51 @@ const LegendCell = styled.td`
   padding: 1px 20px 1px 50px;
 `
 
+const db = firebaseApp.database()
+
 class ParkingStatus extends Component {
   state = {
-    stall: [
-      {ID: 'A1', status: 'vacant', index: 0},
-      {ID: 'A2', status: 'occupied', index: 1},
-      {ID: 'A3', status: 'vacant', index: 2},
-      {ID: 'A4', status: 'occupied', index: 3},
-      {ID: 'A5', status: 'vacant', index: 4},
-      {ID: 'B1', status: 'vacant', index: 5},
-      {ID: 'B2', status: 'vacant', index: 6},
-      {ID: 'B3', status: 'vacant', index: 7},
-      {ID: 'B4', status: 'occupied', index: 8},
-      {ID: 'B5', status: 'reserved', index: 9},
-    ],
+    stalls: []
   }
+
+  componentDidMount() {
+    db.ref('stalls').once('value')
+      .then(snapshot => {
+        let stallsData = snapshot.val()
+        let tmpArr = []
+        for (let stall in stallsData) {
+          tmpArr.push(stallsData[stall])
+        }
+        this.setState({ stalls: tmpArr })
+      }, err => console.log(err))
+  }
+
+  componentDidUpdate() {
+    let copyStalls = this.state.stalls
+    db.ref('stalls').on('child_changed', childSnapshot => {
+      copyStalls[childSnapshot.val().index] = childSnapshot.val()
+      this.setState({ stalls: copyStalls })
+    }, err => console.log(err))
+  }
+
   render(){
-    return(
-      <Wrapper>
-        <Table>
-          <tbody>
-            <Row>
-              <SummaryCell><StatusSummary data={this.state.stall}/></SummaryCell>
-              <StatusCell><StatusTable data={this.state.stall}/></StatusCell>
-              <LegendCell><StatusLegend/></LegendCell>
-            </Row>
-          </tbody>
-        </Table>
-      </Wrapper>
-    )
+    if (this.state.stalls.length) {
+      return (
+        <Wrapper>
+          <Table>
+            <tbody>
+              <Row>
+                <SummaryCell><StatusSummary data={this.state.stalls}/></SummaryCell>
+                <StatusCell><StatusTable data={this.state.stalls}/></StatusCell>
+                <LegendCell><StatusLegend/></LegendCell>
+              </Row>
+            </tbody>
+          </Table>
+        </Wrapper>
+      )
+    }
+
+    return <div>Loading...</div>
   }
 }
