+++ src/StallStatus.js
@@ -19,19 +19,24 @@ const DotYellow = styled(DotGreen)`
 `
 
 const StallStatus = ({status}) => {
-  if (status == 'vacant') {
+  if (status === 'vacant') {
     return (
       <DotGreen></DotGreen>
     )
   }
-  else if (status == 'occupied') {
+  else if (status === 'occupied') {
     return (
       <DotRed></DotRed>
     )
   }
-  else if (status == 'reserved') {
+  else if (status === 'reserved') {
     return (
       <DotYellow></DotYellow>
+    )
+  }
+  else {
+    return (
+      <div>invalid status: {status}</div>
     )
   }
 }
